fix(datagrid): focus trap no longer steals focus when going outside
The focus trap is too greedy for the detail pane use case, where we only want to trap focus for keyboard users. This adds a new local option that puts the rebound elements before and after the host and focus traps only when tabbing through the content. If you use a mouse to move focus outside of the Datagrid, it will disable the focus trap entirely.

fixes #3999

Signed-off-by: Jeremy Wilken <gnomation@gnomeontherun.com>