fix(datagrid): single action overflow close behavior
This change adds details around adding the clrPopoverCloseButton to items projected into the action overflow menu that also need to be able to close the menu. It also explains that the default behavior is to ignore clicks on projected content.
closes #3796

Signed-off-by: Matt Hippely <mhippely@vmware.com>

fix(datagrid): close action overflow content on click

When new popover utility was added to the action overflow the close behavior changed. Previously this menu closed after user clicked on an item in the popover but after the new utility the content stayed open. This change adds a click listener that calls the toggle servcice and toggles it wth the click event. closes #3796

Signed-off-by: Matt Hippely <mhippely@vmware.com>