fix(a11y): type error on focus initialization
The key-focus component does not handle properly the focus initialization, when some of the focusable elements are being removed.
In case the focus was on an element, whose index is no longer available, the component throws a type error.
This change handles the case by initializing the focus on the last focusable element.

Signed-off-by: Zdravko Popov <zpopov@vmware.com>