docs: add form control demos and header demo
Adds demos for the following components:
- header (user can select all variations of light/dark header)
- The general form pattern is to give knob control for: vertical, required, disabled, helpers, errors
- checkbox
- datalist
- input
- input
- password
- radio
- roange
- select
- textarea
- toggle

disabled is not quite right, the observed behavior is that there seems
to be a diconnect between storybook and angular. When value is changed
its not reflected until the control is interacted with (and change
detection is run?).

Cannot get horizontal layout to work. All of these forms are default
layout and _should_ be horizontal. They are not they are vertical.

Related to the last point, for some reason the browser is not following
the sourcemaps to the ts source whcih makes debugging the component
problematic. This is something I'll investigate in a seperate change.

Signed-off-by: Matt Hippely <mhippely@vmware.com>