chore: switch to using yarn
As part of our reorg, we can use yarn to better handle monorepo dependencies. Yarn has workspaces, that will download and manage the dependencies of each package and hoist up duplicates so we can have faster installs and install in one command. This updates only the packages and not the apps.

Signed-off-by: Jeremy Wilken <gnomation@gnomeontherun.com>