feat: enable SHIFT+TAB in focus trap
- if the focus goes out to the top rebound element, it means a user has pressed SHIFT+TAB from the first focusable element in the trap.
So the focus trap will rebound the focus to its last focusable element.
- if the focus goes out to the bottom rebound element, it means a user has pressed TAB from the last focusable element in the trap.
So the focus trap will rebound the focus to its first focusable element.
- if the focus goes out to an element outside that is not rebounder element, it means a user has used a mouse to click and focus.
So the focus trap will rebound the focus to its host element. This is the case if the focus trap config is on strict mode.
If the trap config is on not strict mode, the focus trap will not rebound the focus.

Fixes #2400

Signed-off-by: stsogoo <stsogoo@vmware.com>