fix: remove inline style on wizard-button.ts
The component style of [attr.aria-hidden] in the wizard-button.ts
component is a bit confusing, as its selector doesn't actually
target anything in the component itself. Showing/Hiding the
buttons in the wizard is handled by _wizard.clarity.scss line 353,
and keeping this line here may be a bit confusing.

Signed-off-by: Michael Krotscheck <krotscheck@gmail.com>