fix: header visually goes over the data
When there are two or more words in column title and the column size gets smaller the text starts to wrap. When the text starts to wrap, it forces the header to shrinks vertically and goes over the data in the grid.

The min-height with position:sticky make the .datagrid-header to keep its height and even when the content wraps the height of the .datagrid-header stays the same. That's why the header content goes over the data.

Signed-off-by: Scott Mathis <smathis@vmware.com>