fix: clr-input-wrapper max width
The issue reported was the animated blue underline when an input was given focus was going the full width of the inputs parent container (a .clr-input-wrapper element). This is a partial fix for evergreen browsers. I have an idea that might also work for IE11 but we need to set up the dev app to build for ie11 (e.g compile for es5).

As such, this is a partial fix for #5009.

Signed-off-by: Matt Hippely <mhippely@vmware.com>