fix(datagrid): give all action-item items the menuitem role
Currently the documentation shows adding buttons with this class, but not giving them the `menuitem` role which is required since the wrapping container has a `menu role`.  We could update the documentation to do this, but this approach adds a tiny directive that does it for the user.  Also added a test for this.

Partially fixes:  #4756

Signed-off-by: Alex Mellnik <a.r.mellnik@gmail.com>