feat(modal): reduce change detection cycles for the modal body
This PR reduces change detection cycles for the modal body by replacing
`HostListener` with `Renderer.listen` outside of the zone. Updating `tabindex`
on the host element can be done through the DOM API and that would not require
Angular to run `ApplicationRef.tick()`. Note that the `HostListener` wraps the
actual listener under the hood into the internal Angular function which runs
`markDirty()` before running the actual listener (the decorated class method).

Signed-off-by: Artur Androsovych <arthurandrosovich@gmail.com>