feat(datagrid): reduce change detection cycles for the `clrOutsideClick`
This PR reduces change detection cycles for the `clrOutsideClick` by replacing
`HostListener` with `Renderer.listen` outside of the zone. Basically, it will not
run change detection when the click event happened on the host element, since the
listener will just `return`, it will run `ApplicationRef.tick()` only if the click
event happens outside of the host element. Note that the `HostListener` wraps the
actual listener under the hood into the internal Angular function which runs
`markDirty()` before running the actual listener (the decorated class method).

Signed-off-by: Artur Androsovych <arthurandrosovich@gmail.com>