fix(website): angular alert images for success and info
In Amazon S3 we switched the name of the `success` image with `info` and `info` with `success`.
After little chat with Jeremy we decided that switching the images here will be a faster solution.
We took that decision because we expect the Angular documentation to be gone soon and we will delete the images in S3 for the angular documentation anyway.

Signed-off-by: Bogdan Bogdanov <bbogdanov@vmware.com>