feat(icons): add xs size to icon t-shirt sizes
Consistent with the general size tokens for Clarity core, this change
adds an xs t-shirt size to the icons size api. My use case is in the
cds-navigation component when in vertical layout, not expanded the icon
needs to be about 1/2 the size of the sm icon.

I'm addin ghte change here to keep these changes seperate from the
cds-navigation branch changes.

Signed-off-by: Matt Hippely <mhippely@vmware.com>