fix(tree-view): error EHCAIWC for lazy expandable nodes
Closes #5927

The main reason for EHCAIWC was that [in this function](https://github.com/vmware/clarity/blob/next/packages/angular/projects/clr-angular/src/data/tree-view/models/recursive-tree-node.model.ts#L46) the Promise or the Observable can be resolved too quickly and the `loading` property value can be changed before the next change detection cycle.

I'm resolving the issue with a pattern we used for the [datagrid selection](https://github.com/vmware/clarity/blob/next/packages/angular/projects/clr-angular/src/data/datagrid/providers/selection.ts#L164) that in first case it was producing a EHCAIWC, then we did a fix and the fix was skipping a change detection cycle and we end up with out last solution that seems to be the best one so far.

Signed-off-by: Bogdan Bogdanov <bbogdanov@vmware.com>