feat(datagrid): reduce change detection cycles for `clr-dg-detail`
This PR reduces change detection cycles for the `clr-dg-detail` by replacing
`HostListener` with `Renderer.listen` outside of the zone. We don't need to run
change detections when any keyboard button is pressed. Note that the `HostListener`
wraps the actual listener under the hood into the internal Angular function which runs
`markDirty()` before running the actual listener (the decorated class method).

Signed-off-by: Artur Androsovych <arthurandrosovich@gmail.com>