fix(commitlint): adding another delimiter for scopes
When doing the research we missed that commitlint has
some multi-scope predefined delimiters and what we
are currently using is in conflict with commitlint expectations.
There is a PR in commitlint repo to fix these scenarios by letting
people provide a custom delimiters for scopes but
the owner is not actively working on that.

@bbogdanov will try to join and push that PR to get
in for their next release and unblock us to get back
to what we wanted in the first place.

https://github.com/conventional-changelog/commitlint/pull/2407

Signed-off-by: Bogdan Bogdanov <bbogdanov@vmware.com>