a11y(core:accordion): rename core accordion panel to section
We have a wrapper container that groups an accordion header and its
content. We call this a 'panel' but this leads to a lot of confusion as
the aria- documentation refers to just the content part as 'panel'.
In this change we rename the container to section to avoid this confusion.

Signed-off-by: Jeeyun Lim <jeeyun.lim@gmail.com>