fix(angular:combobox): reduce change detection cycles
This is an adoption of: https://github.com/vmware/clarity/pull/6058

This PR reduces change detection cycles for the combobox by replacing
`HostListener` with `Renderer.listen` outside of the zone. We don't need to
run change detections when any keyboard button is pressed or when the multiselect
mode is off. Note that the `HostListener` wraps the actual listener under the hood
into the internal Angular function which runs `markDirty()` before running the actual
listener (the decorated class method).

Signed-off-by: Artur Androsovych <arthurandrosovich@gmail.com>
Signed-off-by: Matt Hippely <mhippely@vmware.com>