fix(core:button): button border disappears on window resize
I have verified that this is not an issue in vanilla clarity core. This appears to be a browser issue (see PR notes) and it is unclear whether it is exacerbated by react or some other culprit. I tried several different approaches to keep the button borders from disappearing. This includes the unworkable thin keyword as well as other transform settings. This is the only solution that worked consistently.

Signed-off-by: Scott Mathis <smathis@vmware.com>