fix(core:typography): fix override issues with accordion header and content
Content projected into accordion-header was being hard coded to 'secondary'.
The API provided to override it was not being applied. This change removes the specificity
to span content as the template no longer has a span. It defaults the exposed font custom
properties to 'secondary' font-weight and font-size.

Content projected into accordion-content now inherits typography from the global scope.

closes #5915, closes #6347

Signed-off-by: Matt Hippely <mhippely@vmware.com>