refactor(react): fix dev app parcel build issue and refactor to internal-components
There was an issue breaking the dev app build related to panel paths, this change fixes that.
This change re-orgs the internal components into an internal-components directory to reflect
both the folder structure in core and to communicateto consumers that these things are
exposed but internal so they should beware.

Signed-off-by: Matt Hippely <mhippely@vmware.com>