build(angular): update dependencies for angular 13
This also updates some deps to address some github security warnings.
This ended up being a little more than I expected. It turns out angular
is roughly coupled to a few dependencies: typescript and eslint come to
mind in this case. I updated eslint because there was an issue with type
script. I bumped typescript (but not to latest yet) because angular 13
needs 4.4.2. Here is a list of things affected from the change:

- Some of our code was failing the eslint type checks so I updated the
  code or ignored in one case.
- Added an ignore rule becuase angular 13 generates an .angular/cache
  folder. Options were to turn it off or ignore it so I chose to ignore
it.
- remember: another reason for bumping the root level typescript and the
  package/schematic typescript is because the @scehamtics package pulls
in typescript as dep but they do it from a third-party folder inside
their package. This means that the @clr/schematics typescript version
and needs to match that typescript version or the schematics will not
build.
- updated lite-server dep to address gh sec warnings
- updated @commitlint deps to address gh sec warnings
- updated conventional-changelog-cli dep to address gh sec warnings
- updated copy-cli dep to address gh sec warnings

Signed-off-by: Matt Hippely <mhippely@vmware.com>